Move things shared between discv4 and discv5 to p2p package
A lot of things originally written for discv5 are now used by discv4 as
well, so this moves all those things from p2p.discv5 to the p2p package.
