+++ p2p/typing.py
@@ -1,5 +1,4 @@
-from typing import Any, Tuple, TypeVar, Union
-
+from typing import Any, Tuple, TypeVar, Union, NewType, NamedTuple
 
 TCommandPayload = TypeVar('TCommandPayload')
 
@@ -11,3 +10,13 @@ Structure = Union[
 
 Capability = Tuple[str, int]
 Capabilities = Tuple[Capability, ...]
+AES128Key = NewType("AES128Key", bytes)
+Nonce = NewType("Nonce", bytes)
+IDNonce = NewType("IDNonce", bytes)
+NodeID = NewType("NodeID", bytes)
+
+
+class SessionKeys(NamedTuple):
+    encryption_key: AES128Key
+    decryption_key: AES128Key
+    auth_response_key: AES128Key
