+++ tests-trio/p2p-trio/test_enr_db.py
@@ -4,8 +4,8 @@ import pytest
 
 from eth.db.backends.memory import MemoryDB
 
-from p2p.discv5.enr_db import NodeDB
-from p2p.discv5.identity_schemes import (
+from p2p.node_db import NodeDB
+from p2p.identity_schemes import (
     default_identity_scheme_registry,
     IdentitySchemeRegistry,
 )
