+++ tests/core/p2p-proto/test_chain_peer_pool.py
@@ -9,8 +9,8 @@ from rlp import sedes
 from eth.chains.mainnet import MAINNET_GENESIS_HEADER, MAINNET_VM_CONFIGURATION
 from eth.chains.ropsten import ROPSTEN_GENESIS_HEADER, ROPSTEN_VM_CONFIGURATION
 
-from p2p.discv5.constants import IP_V4_ADDRESS_ENR_KEY, TCP_PORT_ENR_KEY, UDP_PORT_ENR_KEY
-from p2p.discv5.typing import NodeID
+from p2p.constants import IP_V4_ADDRESS_ENR_KEY, UDP_PORT_ENR_KEY, TCP_PORT_ENR_KEY
+from p2p.typing import NodeID
 from p2p.kademlia import Node
 from p2p.tools.factories.discovery import ENRFactory
 from p2p.tools.factories.kademlia import IPAddressFactory, NodeFactory
