+++ tests/core/p2p-proto/test_forkid.py
@@ -9,7 +9,7 @@ from eth_utils import to_bytes
 from eth.chains.mainnet import MAINNET_VM_CONFIGURATION
 from eth.chains.ropsten import ROPSTEN_VM_CONFIGURATION
 
-from p2p.discv5.enr import ENR
+from p2p.enr import ENR
 from p2p.exceptions import MalformedMessage
 from p2p.tools.factories.discovery import ENRFactory
