+++ trinity/components/builtin/peer_discovery/component.py
@@ -30,8 +30,8 @@ from p2p.discovery import (
     PreferredNodeDiscoveryService,
     StaticDiscoveryService,
 )
-from p2p.discv5.enr_db import NodeDB
-from p2p.discv5.identity_schemes import default_identity_scheme_registry
+from p2p.node_db import NodeDB
+from p2p.identity_schemes import default_identity_scheme_registry
 
 from trinity.boot_info import BootInfo
 from trinity.config import Eth1AppConfig
